GH-4: Offer a work around via using ssh to communicate with the host
It turns out having nested mounts from the host -> container -> inner
docker container is not so trivial. Other alternatives were to mount
hostpipes and setup a listener on the host machine to exec anything
coming through the pipe.

The Host pipe solution became hard when we wanted to attach to tty's,
and redirecting/formatting of command output was not simple.

An alternative is to mount the ssh keys from the host to allow for a
connection back to the host (a little clunky, however this will serve as
a way to work around the host container issue)

An improvement could be some config to preserve the user creds back to
the host (don't keep requiring passwords for new sessions), or
preferably not at all.

[not really concerned about security of this, this should only be used
as a development environment on a system]

Other changes:

1. Start tmux by default
2. Fix shebang to reflect the bashisms in the install/bootstrap script

Closes #4